+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adds support for page numbers in `.docx` and `.doc` when user or renderer
   created page breaks are present.
+* Adds retry logic for the unstructured-ingest Biomed connector
 
 ### Features